Move perform_db_upgrade, upgrader to sqlite.db_upgrades
and _latest_available_version as well
